[+/*/-]: schedule/tensor/ir/device/dtype/onnx: massive refactoring.
Lands GigaAM ASR end-to-end and brings the IR/runtime/codegen pipeline to
line-for-line parity with tinygrad's CALL/PROGRAM model. The headline
user-visible change is the arena memory planner (default), which packs
disjoint-lifetime buffers into per-(device, copy-lane) arenas and slashes
GigaAM peak memory vs the legacy pool-reuse `Remap` planner. The structural
work underneath is a Buffer redesign, a callable IR (`Op::Call`/`Function`/
`Tuple`/`Program`/`Linear`/`Source`/`ProgramBinary`), and a strict
hazard-partitioned execution plan.

- New ops: `Call`, `Function`, `Tuple`, `GetTuple`, `Program`, `Linear`,
  `Source`, `ProgramBinary`, `CustomFunction`, `LUnique`. Replaces
  `Op::Kernel` and `Op::Assign`; `MovementArg` deleted.
- `Sink::info: Option<KernelInfo>` is now structural and hash-consed (no
  more `KernelAstMarker` type-erased metadata channel).
- Boundary-preserving rewrite drivers (`graph_rewrite_*_preserve_calls`,
  `substitute_preserve_calls`, `graph_rewrite_walk`) keep passes from
  recursing into compiled bodies (`ir/src/rewrite/engine.rs`).
- `ExecutionPlan` rewritten around `PreparedOp { CompiledProgram |
  BufferCopy | BufferView | CustomFunction }` with toposorted dependency
  graph, hazard-partitioned parallel groups by `BufferId` reads/writes,
  symbolic launch sizes, and runtime var-bound validation.
- Legacy `find_independent_kernels`, `can_parallelize`, `KernelAccess`
  partitioning, and unsafe `Send/Sync` impls on `CachedKernel`/
  `PreparedKernel` are removed.

- New `tensor/src/memory_planner/tlsf.rs`: deterministic two-level
  segregated-fit allocator (`block_size=256`, `lv2_cnt=32`, FIFO buckets)
  ported from tinygrad's `runtime/support/memory.py`.
- New `memory_plan_arena` (`tensor/src/memory_planner/mod.rs`): liveness
  analysis splits buffers by `(DeviceSpec, copy-lane?)`, extends
  copy-buffer holds past their last use, runs TLSF over a 2× total-bytes
  budget, sizes one arena per lane to actual peak, and rewrites each
  plannable buffer as a `Buffer::view` into its lane arena.
- `PlannerMode::Arena` is now the default when `MOROK_MEMORY_PLANNER` is
  unset; `0|off|none|disabled` opts out, `remap|pool` selects the legacy
  pool-reuse planner.
- Allocator-overlap on a freed range emits a `ReuseDependency` so the
  runtime preserves anti-dependencies under arena views.

- `Buffer::view` now mints a fresh `BufferId` while sharing the underlying
  `BufferData`. Disjoint arena views appear as independent buffers to the
  hazard model — exactly matching tinygrad's BUFFER_VIEW-as-distinct-
  identity semantics.
- New accessors: `Buffer::storage_id`, `Buffer::total_size`,
  `Buffer::allocator_arc`.
- View-execute assertion compares `storage_id()` instead of handle id
  (`runtime/src/execution_plan.rs`).
- **Migration**: downstream code using `buffer.id()` for storage-equality
  must switch to `storage_id()`.

- New `codegen/src/program_pipeline.rs` enforces `SINK → LINEAR → SOURCE
  → BINARY` staging via `do_linearize`/`do_render`/`do_compile`/
  `get_program`.
- C/LLVM/MLIR renderers now expect `Op::Linear` input and report
  `Error::InvalidGraph` via per-context `pending_error` instead of
  panicking. Multi-index INDEX is rejected at render (consumers run
  `pm_linearize_multi_index` first).
- LLVM text rejects `Custom`/`CustomI` (only C handles them, with a
  strict template formatter); LLVM `lcast` debug-asserts no FP8 inputs.
- `RenderedKernel` slimmed: drops `global_size`/`local_size`/`thread_id`
  plumbing — work-size lives in `ProgramSpec` and is resolved at dispatch.

- **`max_mel_frames` scaling fix** (`model/src/gigaam/mod.rs`): split
  `max_seq_len` into `max_mel_frames` (pre-subsampling JIT input bound)
  and `max_encoder_frames` (post-subsampling/RoPE bound); when configs
  only specify `pos_emb_max_len`, derive `max_mel_frames =
  max_encoder_frames * subsampling_factor`. Audio approaching the encoder
  cap is no longer rejected at the JIT input gate.
- **`reflect_pad` PyTorch parity** (`model/src/audio.rs`): mirrors
  `Reflect1d` (no boundary duplication, single-bounce only with
  `pad < len`).
- **CTC head dtype** (`model/src/gigaam/head.rs`): upcasts log_softmax
  inputs to fp32 only when activation is sub-fp32 (replaces incorrect
  downcast to weight dtype).
- **Encoder dtype/RoPE** (`model/src/gigaam/encoder.rs`): caches
  `d_model` on `MultiHeadSelfAttention`, materializes RoPE once at
  `[B,T,d_model]` so Q/K share a single rotated buffer; re-casts BN/LN
  output back to activation dtype to preserve precision.
- **JIT improvements** (`model/src/jit.rs`): adds `DuplicateInputBuffer`
  error and boxes `morok_tensor::Error` to shrink variant size.
- **`gigaam_infer.rs` example**: dedupes buffer summary by `storage_id`,
  prints launch dims symbolically, unwraps kernel ASTs from
  `Op::Call { body }`.

- **Pre/post-schedule cache split** (`tensor/src/realize.rs` + new
  `schedule_cache.rs` shape): cached entry stores a `PreSchedule` of
  callable items rather than a kernelized `UOp`;
  `restore_bind_placeholders_for_schedule` re-injects BIND only before
  rangeify so the kernel AST stays parametric across runs.
- **OptCache FIFO eviction**: `OptCacheState` bounds the global
  optimize+compile cache via `MOROK_OPT_CACHE_MAX` (default 4096).
- **Schedule cache disable** via `PrepareConfig.disable_schedule_cache`
  or `MOROK_DISABLE_SCHEDULE_CACHE=1`.
- **Side-realize removal**: `resolve_pending_assigns` is gone;
  `Tensor::try_assign` embeds writes as `AFTER(target, STORE(target,
  value))` directly.
- **`isinf` correctness fix** (`tensor/src/math.rs`): bitcast-and-compare
  against ±inf bit patterns instead of `(x-x).isnan()` (which Morok's
  range analyzer folds away).

- `MOROK_BEAM_MIN_PROGRESS` epsilon termination (default 10 ns), mirroring
  tinygrad `BEAM_MIN_PROGRESS` (`schedule/src/optimizer/beam.rs`).
- TC search caps axis-choice retries.
- `pm_mod_to_idiv` lowers `x % d → x - d*(x//d)` for non-pow2 const
  divisors so fast_idiv fires on the resulting IDIV.
- z3 `Cast` binds equality `cast == src` only when source bounds fit
  destination dtype (avoids spurious UNSAT under narrowing casts that
  would falsely "verify" arbitrary equivalences).

- `Program` reference-counted (`Arc<dyn Program>`) and moved from `queue`
  to `device`; `Send + Sync` required.
- `ProgramSpec` switches `global_size`/`local_size` from concrete
  `Option<[usize; 3]>` to symbolic `[Arc<UOp>; 3]`, resolved per dispatch
  via `resolve_launch_dims`. Special-axis names `g{N}`/`l{N}`/`i{N}`
  drive launch derivation.
- `CudaTimelineSignal` replaces the `HashMap<u64, Arc<CudaEvent>>` event
  pool with a fixed 64-slot ring (`EventRing`), fixing torn-waiter races
  on ring wrap.

- 18 new whitebox `*_internal.rs` modules across `ir`, `schedule`,
  `tensor`. Convention: source module ends with `#[cfg(test)] #[path =
  "../test/unit/.../<name>_internal.rs"] mod tests;` so the test file
  gains `use super::*;` access to private items, while public-API tests
  stay in the original `<name>.rs` files.
- New blackbox test files: `tensor/src/test/unit/{custom_kernel,
  schedule, schedule_cache, tensor_registry, tlsf}.rs`,
  `device/src/test/unit/sync.rs`, `model/src/test/unit/audio.rs`,
  ~9 new `runtime/src/test/unit/*.rs` files.
- `.tokeignore` tracks the new test trees.

+++ model/src/test/unit/mel.rs
@@ -69,8 +69,8 @@ fn test_mel_spectrogram_sine_wave() {
 
     let vals = output.as_vec::<f32>().unwrap();
     let shape = output.shape().unwrap();
-    let n_mels = shape[1].as_const().unwrap() as usize;
-    let n_frames = shape[2].as_const().unwrap() as usize;
+    let n_mels = shape[1].as_const().unwrap();
+    let n_frames = shape[2].as_const().unwrap();
 
     let mut avg_energy: Vec<f32> = vec![0.0; n_mels];
     for mel_idx in 0..n_mels {
